requirements.txt
pandas==1.2.2
tensorflow==2.4.1
Pillow==8.1.0
tqdm==4.56.0
